Moved UI related to selecting Game to GamesMenu from StatsMenu. Hooked up GamesMenu to MenuContainer.
